Reorganize pictures
As new pictures were added, they were put in a new Final_Pics
directory. But as things progressed, some of those became unused
while some of the original pics were used. So the needed pictures
were a mix of the two directores, each of which contained lots
of unused imageas. Reorganized so that all used pictures are
in `pics` and the unused ones are in `pics-old`.